fix: 모바일팀에서 baseOnBalls에 대한 요구한 내용 float에서 integer로 적용

+++ be/server/src/main/java/com/baseball/domain/player/Pitcher.java
@@ -27,8 +27,8 @@ public class Pitcher extends Player {
         return numberOfPitching;
     }
 
-    public Float getBaseOnBalls() {
-        return (float) ballCount / 4;
+    public Integer getBaseOnBalls() {
+        return ballCount / 4;
     }
 
     private Integer getOutCount() {
